move trigger sequence to stage-overview
