+++ bridge/client/app/_views/ktb-environment-view/ktb-stage-overview/ktb-stage-overview.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { KtbStageOverviewComponent } from './ktb-stage-overview.component';
 import { DtFilterFieldModule } from '@dynatrace/barista-components/filter-field';
 import { FlexModule } from '@angular/flex-layout';
-import { KtbTriggerSequenceModule } from '../../../_components/ktb-trigger-sequence/ktb-trigger-sequence.module';
 import { KtbSelectableTileModule } from '../../../_components/ktb-selectable-tile/ktb-selectable-tile.module';
 import { DtIconModule } from '@dynatrace/barista-components/icon';
 import { DtButtonModule } from '@dynatrace/barista-components/button';
@@ -16,26 +15,38 @@ import { DtOverlayModule } from '@dynatrace/barista-components/overlay';
 import { KtbEvaluationInfoModule } from '../../../_components/ktb-evaluation-info/ktb-evaluation-info.module';
 import { DtShowMoreModule } from '@dynatrace/barista-components/show-more';
 import { KtbNoServiceInfoModule } from '../../../_components/ktb-no-service-info/ktb-no-service-info.module';
+import { DtFormFieldModule } from '@dynatrace/barista-components/form-field';
+import { DtInputModule } from '@dynatrace/barista-components/input';
+import { DtRadioModule } from '@dynatrace/barista-components/radio';
+import { DtSelectModule } from '@dynatrace/barista-components/select';
+import { FormsModule } from '@angular/forms';
+import { KtbDateInputModule } from '../../../_components/ktb-date-input/ktb-date-input.module';
+import { KtbTriggerSequenceComponent } from './ktb-trigger-sequence/ktb-trigger-sequence.component';
 
 @NgModule({
-  declarations: [KtbStageOverviewComponent, KtbServicesListComponent],
+  declarations: [KtbStageOverviewComponent, KtbServicesListComponent, KtbTriggerSequenceComponent],
   imports: [
     CommonModule,
     DtButtonModule,
     DtFilterFieldModule,
+    DtFormFieldModule,
     DtIconModule.forRoot({
       svgIconLocation: `assets/icons/{{name}}.svg`,
     }),
+    DtInputModule,
     DtOverlayModule,
+    DtRadioModule,
+    DtSelectModule,
     DtShowMoreModule,
     DtTableModule,
     FlexModule,
+    FormsModule,
+    KtbDateInputModule,
     KtbEvaluationInfoModule,
     KtbLoadingModule,
     KtbNoServiceInfoModule,
     KtbPipeModule,
     KtbSelectableTileModule,
-    KtbTriggerSequenceModule,
     RouterModule,
   ],
   exports: [KtbStageOverviewComponent],
