+++ bridge/client/app/_views/ktb-environment-view/ktb-stage-overview/ktb-trigger-sequence/ktb-trigger-sequence.component.html
