+++ bridge/client/app/_views/ktb-environment-view/ktb-stage-overview/ktb-trigger-sequence/ktb-trigger-sequence.component.scss
@@ -1,5 +1,5 @@
-@import '../../../variables';
-@import '~@dynatrace/barista-components/style/main';
+@import 'client/variables';
+@import 'node_modules/@dynatrace/barista-components/style/main';
 
 .ktb-trigger-sequence {
   background-color: $gray-100;
